+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Updated
-  * Updated Without Default Time example to have visual label in `terra-time-input`.
+  * Updated `terra-time-input` example to have visual label.
 
 * Changed
   * Updates `terra-theme-provider` examples to show the new ability to specify a theme density.