+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added implementation guide for `terra-table`.
+  
 * Changed
   * Updates `terra-theme-provider` examples to show the new ability to specify a theme density.